Rename issuance_details.rb to attribute_issuance_details.rb
