+++ lib/yoti/share/attribute_issuance_details.rb
